project-5/pyproject.toml
[tool.poetry]
name = "project-5"
version = "0.1.0"
description = "Project-5 Integrate Postgres Database & create Table"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.3"
gspread = "^5.12.0"
gspread-dataframe = "^3.3.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
